+++ src/main/java/io/wiretap/kafka/message/settings/SpecificKafkaInfoLogMessageSettings.java
@@ -29,6 +29,10 @@ public class SpecificKafkaInfoLogMessageSettings extends KafkaInfoLogMessageSett
         merged.setHeaders(
                 CollectionUtils.isEmpty(this.getHeaders()) ? common.getHeaders() : this.getHeaders()
         );
+        merged.setKeyJsonInclude(
+                CollectionUtils.isEmpty(this.getKeyJsonInclude())
+                        ? common.getKeyJsonInclude() : this.getKeyJsonInclude()
+        );
         merged.setVisibilitySettings(
                 this.getVisibilitySettings().equals(defaults.getVisibilitySettings())
                         ? common.getVisibilitySettings() : this.getVisibilitySettings()